# frontend/openapi.json
{
   "openapi": "3.1.0",
   "info": {
      "title": "Public API",
      "version": "1.0"
   },
   "paths": {
...
      "/api/internal/auth/refresh": {
         "post": {
            "tags": [
               "auth"
            ],
            "summary": "Refresh",
            "operationId": "Refresh Token",
            "requestBody": {
               "content": {
                  "application/json": {
                     "schema": {
                        "$ref": "#/components/schemas/RefreshRequest"
                     }
                  }
               },
               "required": true
            },
            "responses": {
               "200": {
                  "description": "Successful Response",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/RefreshResponse"
                        }
                     }
                  }
               },
               "422": {
                  "description": "Validation Error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      },
...
      "/api/internal/account/{account_id}/": {
         "get": {
            "summary": "Get User",
            "operationId": "Get Account Details",
            "security": [
               {
                  "OAuth2PasswordBearer": []
               }
            ],
            "parameters": [
               {
                  "name": "account_id",
                  "in": "path",
                  "required": true,
                  "schema": {
                     "type": "string",
                     "format": "uuid",
                     "title": "Account Id"
                  }
               }
            ],
            "responses": {
               "200": {
                  "description": "Successful Response",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/AccountDetailsResponse"
                        }
                     }
                  }
               },
               "422": {
                  "description": "Validation Error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/HTTPValidationError"
                        }
                     }
                  }
               }
            }
         }
      },
...
   },
   "components": {
      "schemas": {
         "AccountDetailsResponse": {
            "properties": {
               "user": {
                  "$ref": "#/components/schemas/UserDetailsResponse"
               },
               "admin": {
                  "anyOf": [
                     {
                        "$ref": "#/components/schemas/AdminDetailsResponse"
                     },
                     {
                        "type": "null"
                     }
                  ]
               },
               "customer": {
                  "anyOf": [
                     {
                        "$ref": "#/components/schemas/CustomerDetailsResponse"
                     },
                     {
                        "type": "null"
                     }
                  ]
               }
            },
            "type": "object",
            "required": [
               "user",
               "admin",
               "customer"
            ],
            "title": "AccountDetailsResponse"
         },
         "AdminDetailsResponse": {
            "properties": {
               "id": {
                  "type": "string",
                  "format": "uuid",
                  "title": "Id"
               },
               "promoted_by_id": {
                  "anyOf": [
                     {
                        "type": "string",
                        "format": "uuid"
                     },
                     {
                        "type": "null"
                     }
                  ],
                  "title": "Promoted By Id"
               }
            },
            "type": "object",
            "required": [
               "id",
               "promoted_by_id"
            ],
            "title": "AdminDetailsResponse"
         },
...
         "CustomerDetailsResponse": {
            "properties": {
               "id": {
                  "type": "string",
                  "format": "uuid",
                  "title": "Id"
               },
               "funds": {
                  "type": "number",
                  "title": "Funds"
               }
            },
            "type": "object",
            "required": [
               "id",
               "funds"
            ],
            "title": "CustomerDetailsResponse"
         },
...
         "HTTPValidationError": {
            "properties": {
               "detail": {
                  "items": {
                     "$ref": "#/components/schemas/ValidationError"
                  },
                  "type": "array",
                  "title": "Detail"
               }
            },
            "type": "object",
            "title": "HTTPValidationError"
         },
...
         "RefreshRequest": {
            "properties": {
               "user_id": {
                  "type": "string",
                  "format": "uuid",
                  "title": "User Id"
               },
               "session": {
                  "type": "string",
                  "title": "Session"
               }
            },
            "type": "object",
            "required": [
               "user_id",
               "session"
            ],
            "title": "RefreshRequest"
         },
         "RefreshResponse": {
            "properties": {
               "token_type": {
                  "type": "string",
                  "const": "bearer",
                  "title": "Token Type",
                  "default": "bearer"
               },
               "access_token": {
                  "type": "string",
                  "title": "Access Token"
               },
               "data": {
                  "$ref": "#/components/schemas/TokenData"
               }
            },
            "type": "object",
            "required": [
               "access_token",
               "data"
            ],
            "title": "RefreshResponse"
         },
         "TokenData": {
            "properties": {
               "id": {
                  "type": "string",
                  "format": "uuid",
                  "title": "Id"
               },
               "email": {
                  "type": "string",
                  "title": "Email"
               },
               "role": {
                  "items": {
                     "$ref": "#/components/schemas/UserRole"
                  },
                  "type": "array",
                  "uniqueItems": true,
                  "title": "Role"
               }
            },
            "type": "object",
            "required": [
               "id",
               "email",
               "role"
            ],
            "title": "TokenData"
         },
         "UserDetailsResponse": {
            "properties": {
               "id": {
                  "type": "string",
                  "format": "uuid",
                  "title": "Id"
               },
               "email": {
                  "type": "string",
                  "title": "Email"
               },
               "email_verified": {
                  "type": "boolean",
                  "title": "Email Verified"
               },
               "role": {
                  "items": {
                     "$ref": "#/components/schemas/UserRole"
                  },
                  "type": "array",
                  "uniqueItems": true,
                  "title": "Role"
               },
               "created_at": {
                  "type": "string",
                  "format": "date-time",
                  "title": "Created At"
               }
            },
            "type": "object",
            "required": [
               "id",
               "email",
               "email_verified",
               "role",
               "created_at"
            ],
            "title": "UserDetailsResponse"
         },
...
         "UserRole": {
            "type": "string",
            "enum": [
               "CUSTOMER",
               "ADMIN"
            ],
            "title": "UserRole"
         },
...
         "ValidationError": {
            "properties": {
               "loc": {
                  "items": {
                     "anyOf": [
                        {
                           "type": "string"
                        },
                        {
                           "type": "integer"
                        }
                     ]
                  },
                  "type": "array",
                  "title": "Location"
               },
               "msg": {
                  "type": "string",
                  "title": "Message"
               },
               "type": {
                  "type": "string",
                  "title": "Error Type"
               },
               "input": {
                  "title": "Input"
               },
               "ctx": {
                  "type": "object",
                  "title": "Context"
               }
            },
            "type": "object",
            "required": [
               "loc",
               "msg",
               "type"
            ],
            "title": "ValidationError"
         },
...
      },
      "securitySchemes": {
         "OAuth2PasswordBearer": {
            "type": "oauth2",
            "flows": {
               "password": {
                  "scopes": {},
                  "tokenUrl": "/api/internal/auth/login"
               }
            }
         },
         "APIKeyHeader": {
            "type": "apiKey",
            "in": "header",
            "name": "X-API-Key"
         }
      }
   }
}